# massive performance improvement for blacklist filtering