--- a/src/mediathek/daten/ListeBlacklist.java
+++ b/src/mediathek/daten/ListeBlacklist.java
@@ -22,6 +22,9 @@ package mediathek.daten;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.LinkedList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import mSearch.daten.DatenFilm;
 import mSearch.daten.ListeFilme;
 import mSearch.tool.Duration;
@@ -46,8 +49,7 @@ public class ListeBlacklist extends LinkedList<DatenBlacklist> {
     public synchronized boolean addProgStart(DatenBlacklist b) {
         //nur beim Programmstart!!, wird nicht gemeldet
         b.arr[DatenBlacklist.BLACKLIST_NR] = getNr(nr++);
-        boolean ret = super.add(b);
-        return ret;
+        return super.add(b);
     }
 
     @Override
@@ -134,16 +136,21 @@ public class ListeBlacklist extends LinkedList<DatenBlacklist> {
 //                }
 //            });
 //            Duration.counterStop("Black-alt");
-//            Duration.counterStart("Black-neu");
-            this.stream().forEach(bl -> bl.toLower());
-            this.stream().forEach(bl -> bl.hasPattern());
-            listeFilme.stream().filter(this::checkFilm_).forEach(filmEntry -> {
-                listeRet.add(filmEntry);
-                if (filmEntry.isNew()) {
-                    listeRet.neueFilme = true;
-                }
+//            listeRet.clear();
+
+            this.parallelStream().forEach(entry -> {
+                entry.toLower();
+                entry.hasPattern();
             });
-//            Duration.counterStop("Black-neu");
+            listeRet.neueFilme = false;
+
+//            Duration.counterStart("Black-neu parallel");
+
+            final List<DatenFilm> col = listeFilme.parallelStream().filter(this::checkFilm_).collect(Collectors.toList());
+            col.parallelStream().filter(DatenFilm::isNew).findFirst().ifPresent(ignored -> listeRet.neueFilme = true);
+            listeRet.addAll(col);
+//            Duration.counterStop("Black-neu parallel");
+            col.clear();
 
             // Array mit Sendernamen/Themen füllen
             listeRet.themenLaden();