Sehr gute Leistungssteigerung bei Anwendung einer Blacklist.
Im Schnitt 100% Leistungssteigerung :-)
Folgende Werte wurden vergleichsweise auf den unterstützten Plattformen ermittelt:

Linux (dual core):
#################################################################
Black-alt                   Anzahl: 1   Gesamtdauer: 3.00 s
Black-neu parallel          Anzahl: 1   Gesamtdauer: 1.29 s
#################################################################

OS X (quad core):
========== ========== ========== ========== ==========
DURATION 8:  Black-alt  [2,29 s]
   Klasse:  ListeBlacklist.filterListe
   Black-alt Anzahl: 1   Dauer: 2,29 s
========== ========== ========== ========== ==========
========== ========== ========== ========== ==========
DURATION 9:  Black-neu parallel  [788,00 ms]
   Klasse:  ListeBlacklist.filterListe
   Black-neu parallel Anzahl: 1   Dauer: 787,00 ms
========== ========== ========== ========== ==========

Windows 10 (dual core):
#################################################################
Black-alt                   Anzahl: 1   Gesamtdauer: 3,62 s
Black-neu parallel          Anzahl: 1   Gesamtdauer: 1,52 s
#################################################################