@claell 2016-09-02
Hört sich gut an. Ich habe erst vor ein paar Monaten die Blacklist entdeckt und mittlerweile mit vielen Einträgen (vermutlich über 100) merkt man deutlich Performance-Einbußen, auch z.B. beim Hinzufügen zur Blacklist.